fix/break import cycles (#46)
- break import cycle between settings<->identity modules
- factor out Sink enum into types.py file
- upgrade versions
