+++ .pre-commit-config.yaml
@@ -3,7 +3,7 @@ default_language_version:
   python: python3.11
 repos:
   - repo: https://github.com/astral-sh/ruff-pre-commit
-    rev: v0.1.3
+    rev: v0.1.4
     hooks:
       - id: ruff
         args: [--fix, --exit-non-zero-on-fix]
@@ -32,7 +32,7 @@ repos:
           - "backoff>=2.2.1,<3"
           - "click>=8.1.7,<9"
           - "pydantic[dotenv,email]>=1.10.12,<2"
-          - "pytest>=7.4.2,<8"
+          - "pytest>=7.4.3,<8"
           - "types-pytz>=2023.3.1.1,<2024"
           - "types-requests>=2.31.0.8,<3"
           - "types-setuptools>=68.2.0.0,<69"
