+++ mex/common/sinks/__init__.py
@@ -1,10 +1,3 @@
-from enum import Enum
+from mex.common.sinks.types import Sink
 
-
-class Sink(Enum):
-    """Configuration to choose where to send outbound data."""
-
-    BACKEND = "backend"
-    GRAPH = "graph"
-    NDJSON = "ndjson"
-    PUBLIC = "public"
+__all__ = ("Sink",)
