+++ mex/common/sinks/types.py
@@ -1,0 +1,10 @@
+from enum import Enum
+
+
+class Sink(Enum):
+    """Configuration to choose where to send outbound data."""
+
+    BACKEND = "backend"
+    GRAPH = "graph"
+    NDJSON = "ndjson"
+    PUBLIC = "public"
