+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "mex-common"
-version = "0.14.0"
+version = "0.14.1"
 description = "RKI MEx common library."
 authors = ["RKI MEx Team <mex@rki.de>"]
 readme = "README.md"
@@ -24,9 +24,9 @@ black = "^23.10.1"
 ipdb = "^0.13.13"
 mypy = "^1.6.1"
 mypy-extensions = "^1.0.0"
-pytest = "^7.4.2"
+pytest = "^7.4.3"
 pytest-cov = "^4.1.0"
-ruff = "^0.1.3"
+ruff = "^0.1.4"
 sphinx = "^6.2.1"
 sphinx-book-theme = "^1.0.1"
 types-pytz = "^2023.3.1.1"
@@ -87,5 +87,5 @@ known-first-party = ["mex", "tests"]
 convention = "google"
 
 [build-system]
-requires = ["poetry-core>=1.7.0"]
+requires = ["poetry-core>=1.8.1"]
 build-backend = "poetry.core.masonry.api"
