+++ requirements.txt
@@ -1,3 +1,3 @@
-poetry==1.6.1
+poetry==1.7.0
 pre-commit==3.5.0
-wheel==0.41.2
+wheel==0.41.3
